+++ public/Stop-SynapseTriggers.ps1
@@ -8,13 +8,9 @@ Stops (disables) all triggers in Synapse Workspace instance (service).
 .PARAMETER SynapseWorkspace
 Name of Synapse Workspace service to be affected.
 
-.PARAMETER ResourceGroupName
-Resource Group Name of Synapse Workspace service to be affected.
-
 .EXAMPLE
-$ResourceGroupName = 'rg-devops'
 $SynapseWorkspace = "SQLPlayerSynapseDemo"
-Stop-SynapseTriggers -SynapseWorkspace "$SynapseWorkspace" -ResourceGroupName "$ResourceGroupName"
+Stop-SynapseTriggers -SynapseWorkspace "$SynapseWorkspace"
 
 .LINK
 Online version: https://github.com/SQLPlayer/azure.synapse.tools/
@@ -22,14 +18,13 @@ Online version: https://github.com/SQLPlayer/azure.synapse.tools/
 function Stop-SynapseTriggers {
     [CmdletBinding()]
     param (
-        [parameter(Mandatory = $true)] [String] $SynapseWorkspace,
-        [parameter(Mandatory = $true)] [String] $ResourceGroupName
+        [parameter(Mandatory = $true)] [String] $SynapseWorkspace
     )
 
     [Synapse] $synapse = New-Object 'Synapse'
     $synapse.Name = $SynapseWorkspace
-    $synapse.ResourceGroupName = $ResourceGroupName
 
+    $synapse.PublishOptions = New-SynapsePublishOption
     Stop-Triggers -synapse $synapse
     
 }