+++ changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Replacing all properties environment-related fails
 
+## [0.18.0] - 2022-01-20
+* Fixed Stop/Start triggers while publishing (#8)
+* Added Start-SynapseTriggers cmdlet
+
 ## [0.17.0] - 2022-01-19
 * Added cmdlet generates dependencies diagram (Get-SynapseDocDiagram)
 * Fixed issue #7: Replacing all properties environment-related fails